Do not pickle node_class, create the node on the fly
