+++ node_graph/orm/mapping.py
@@ -1,0 +1,7 @@
+type_mapping = {
+    "default": "node_graph.any",
+    int: "node_graph.int",
+    float: "node_graph.float",
+    str: "node_graph.string",
+    bool: "node_graph.bool",
+}
